backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
google-generativeai==0.3.2
reportlab==4.0.7
python-docx==1.1.0
Pillow==10.1.0
PyPDF2==3.0.1
python-magic==0.4.27
gitpython==3.1.40
beautifulsoup4==4.12.2
lxml==4.9.3